Update to v097r05 release.
byuu says:

More V30MZ implemented, a lot more to go.

icarus now supports importing WS and WSC games. It expects them to have
the correct file extension, same for GB and GBC.

> Ugh, apparently HiDPI icarus doesn't let you press the check boxes.

I set the flag value in the plist to false for now. Forgot to do it for
higan, but hopefully I won't forget before release.

+++ higan/ws/cpu/cpu.hpp
@@ -7,8 +7,8 @@ struct CPU : Processor::V30MZ, Thread, IO {
   auto wait(uint clocks = 1) -> void override;
   auto read(uint20 addr) -> uint8 override;
   auto write(uint20 addr, uint8 data) -> void override;
-  auto in(uint16 port) -> uint16 override;
-  auto out(uint16 port, uint16 data) -> void override;
+  auto in(uint16 port) -> uint8 override;
+  auto out(uint16 port, uint8 data) -> void override;
 
   auto power() -> void;
 
@@ -16,8 +16,8 @@ struct CPU : Processor::V30MZ, Thread, IO {
   auto ramRead(uint16 addr) -> uint8;
   auto ramWrite(uint16 addr, uint8 data) -> void;
 
-  auto portRead(uint16 addr) -> uint16 override;
-  auto portWrite(uint16 addr, uint16 data) -> void override;
+  auto portRead(uint16 addr) -> uint8 override;
+  auto portWrite(uint16 addr, uint8 data) -> void override;
 };
 
 extern CPU cpu;
